vscroll 1.0.0.rc-8 + adapter.clip with "infinite" spec
