+++ package-dist.json
@@ -33,7 +33,7 @@
   },
   "dependencies": {
     "tslib": "^1.9.0",
-    "vscroll": ">=1.0.0-rc.7 <2.0.0"
+    "vscroll": ">=1.0.0-rc.8 <2.0.0"
   },
   "keywords": [
     "angular",
